characters for the score, character drawing
The area covered by the blinds must be repainted fully every frame,
because the screen is scrolling.

It is important that it's static and clean. Unfortunately the push
method of memory filling is not great in this case because it produces
artifacts at a frame boundary.

Only the alpha layer needs to be updated with full character bitmaps.
The actual white is just filled with ff. It's possible that some of this
could be optimised to use push-filling.