game progression
Bounce flag is now used as frozen flag, so 5 bytes are copied back after
a foe is processed. A pretty sad slowdown which can be helped by some
moving around of the fields in the foe structure.

Changes that contribute to the feeling that the game becomes harder
from bridge to bridge.