fix level repeatability
Game state is saved when crossing the bridge, this includes
frame_scroll. frame_scroll however does not get updated after game
restore (which probably should be fixed) and instead an increment is
added to GameResetToBridge is added.

This happens because some things in CreateNewFoe depend on frame_scroll
value, which is also probably not ideal but that's just how it is.

Added fancy debugging stuff via i/o ports.